lat/lng fixed

+++ backend/db/migrations/20230514155014-create-spot.js
@@ -41,11 +41,11 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.FLOAT(8,6),
+        type: Sequelize.DECIMAL(8,6),
         allowNull: false
       },
       lng: {
-        type: Sequelize.FLOAT (9,6),
+        type: Sequelize.DECIMAL (9,6),
         allowNull: false,
       },
       name: {
